Idea - Create a test "result" as migration_notices instance attribute, and bubble it upwards to where things are far more conducive to doing I/O.
It means the developer gets the feedback at the end of running tests, but that's probably a better place in the grand scheme of things because full test runs can take a very long time, so the most important info tends to run towards the end of the test output where this ends up. Open to other ideas if they demonstrate a usability improvement.